iter4 Разбила общий хэндлер на несколько
