+++ internal/handlers/set_short_url/handler_test.go
@@ -1,4 +1,4 @@
-package handler
+package set_short_url
 
 import (
 	"io"
@@ -11,56 +11,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestHandler(t *testing.T) {
-	// Получаем короткую ссылку для URL
-	url := "https://practicum.yandex.ru"
-	r := NewRouter()
-	ts := httptest.NewServer(r)
-
-	statusCode, body := testRequest(t, ts, http.MethodPost, "/", strings.NewReader(url))
-	assert.Equal(t, http.StatusCreated, statusCode)
-	assert.Equal(t, "http://localhost:8080/6bdb5b0e26a76e4dab7cd1a272caebc0", body)
-
-	// По короткой ссылке получаем полный URL
-	statusCode, _ = testRequest(t, ts, http.MethodGet, "/6bdb5b0e26a76e4dab7cd1a272caebc0", nil)
-	assert.Equal(t, http.StatusOK, statusCode)
-	ts.Close()
-}
-
-func TestHandlerGetNegative(t *testing.T) {
-	type want struct {
-		statusCode int
-		response   string
-	}
-	tests := []struct {
-		name    string
-		request string
-		method  string
-		want    want
-	}{
-		{
-			name:    "GET - URL по заданному ID не найден",
-			request: "/123",
-			method:  http.MethodGet,
-			want: want{
-				statusCode: http.StatusBadRequest,
-				response:   "not found\n",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := NewRouter()
-			ts := httptest.NewServer(r)
-			defer ts.Close()
-
-			statusCode, body := testRequest(t, ts, tt.method, tt.request, nil)
-			assert.Equal(t, tt.want.statusCode, statusCode)
-			assert.Equal(t, tt.want.response, body)
-		})
-	}
-}
 
 func TestHandlerPostNegative(t *testing.T) {
 	type want struct {
@@ -123,7 +73,6 @@ func NewRouter() chi.Router {
 	r := chi.NewRouter()
 
 	r.Route("/", func(r chi.Router) {
-		r.Get("/{id}", Handler)
 		r.Post("/", Handler)
 	})
 	return r
